pyproject.toml
[project]
name = "pdf-info-gather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=5.14.0",
    "groq>=0.16.0",
    "ipykernel>=6.29.5",
    "modelscope-studio>=1.0.2",
    "openai>=1.61.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "tqdm>=4.67.1",
]
